adicionando os ataques nos guerreiros

+++ trabalhoPOO/src/guerreiros/egipcios/tipos/MortoVivo.java
@@ -3,7 +3,7 @@ package guerreiros.egipcios.tipos;
 import guerreiros.egipcios.Egipcios;
 
 public class MortoVivo extends Egipcios {
-    public MortoVivo(String nome, Integer idade, Double peso) {
+    public MortoVivo(String nome, Integer idade, Integer peso) {
         super(nome, idade, peso);
     }
 }
